+++ src/main/java/spot/spot/global/response/format/GlobalException.java
@@ -1,4 +1,4 @@
-package spot.spot.global.exception;
+package spot.spot.global.response.format;
 
 import lombok.Getter;
