+++ src/main/java/spot/spot/global/response/handler/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package spot.spot.global.exception;
+package spot.spot.global.response.handler;
 
 
 import lombok.RequiredArgsConstructor;
@@ -7,7 +7,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import spot.spot.global.dto.response.ResultResponse;
+import spot.spot.global.response.format.ResultResponse;
+import spot.spot.global.response.format.GlobalException;
 
 @RestControllerAdvice
 @RequiredArgsConstructor
