+++ src/test/java/spot/spot/global/config/ConnectToKlaytnNetworkTest.java
@@ -9,7 +9,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import spot.spot.global.response.handler.ConnectToKlaytnNetwork;
+import spot.spot.global.klaytn.ConnectToKlaytnNetwork;
 
 import java.math.BigInteger;
